pyproject.toml
[tool.poetry]
name = "drf-project"
version = "0.1.0"
description = ""
authors = ["kot <bafomet2016vitt@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.5"
Pillow = "^10.0.1"
django-rest-framework = "^0.1.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.8"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
